pyproject.toml
[project]
name = "hf-agents-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=0.31.1",
    "litellm>=1.68.1",
    "python-dotenv>=1.1.0",
    "smolagents[all,litellm]>=1.15.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "wheel>=0.45.1",
]
